test(func): move timeout test in a dedicated file
